instances: move code to separate package
Since 'list' code was moved to binary package, listInstances should be
moved to separate package.

Follow up #628
